Separate code in new file Consider statements in actions
