+++ testdata/p4_16_samples_outputs/pna-action-selector.p4-error
@@ -1,0 +1,20 @@
+mov h.ethernet.dstAddr; t.param;
+mov h.ethernet.etherType; t.param;
+mov m.MainControlT_as_group_id; t.group_id;
+<DpdkRxStatement>(390414)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+mov m.MainControlT_as_member_id; 0
+mov m.MainControlT_as_group_id; 0
+apply tbl
+apply as_sel
+apply as
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(390420)
