+++ testdata/p4_16_samples_outputs/pna-action-selector.p4.spec
@@ -43,7 +43,7 @@ metadata instanceof main_metadata_t
 header ethernet instanceof ethernet_t
 header ipv4 instanceof ipv4_t
 
-regarray direction_port_mask size 0x1 initval 0
+regarray network_port_mask size 0x1 initval 0
 
 action NoAction args none {
 	return
