+++ testdata/p4_16_samples_outputs/pna-direction.p4-error
@@ -1,0 +1,41 @@
+mov m.pna_main_output_metadata_output_port; t.vport;
+drop
+<DpdkRxStatement>(385797)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+jmpneq LABEL_TRUE m.pna_pre_input_metadata_direction; 0
+LABEL_FALSE
+mov m.local_metadata_b; 0
+jmp LABEL_END
+LABEL_TRUE
+mov m.local_metadata_b; 1
+LABEL_END
+jmpneq LABEL_FALSE_0 0 m.pna_pre_input_metadata_direction;
+LABEL_TRUE_0
+mov m.local_metadata_tmpDir; h.ipv4.srcAddr;
+jmp LABEL_END_0
+LABEL_FALSE_0
+mov m.local_metadata_tmpDir; h.ipv4.dstAddr;
+LABEL_END_0
+jmpneq LABEL_FALSE_1 0 m.pna_main_input_metadata_direction;
+LABEL_TRUE_1
+mov m.MainControlT_tmpDir; h.ipv4.srcAddr;
+jmp LABEL_END_1
+LABEL_FALSE_1
+mov m.MainControlT_tmpDir; h.ipv4.dstAddr;
+LABEL_END_1
+jmpnv LABEL_FALSE_2 h.ipv4;
+LABEL_TRUE_2
+apply ipv4_da_lpm
+jmp LABEL_END_2
+LABEL_FALSE_2
+LABEL_END_2
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(385803)
