+++ testdata/p4_16_samples_outputs/pna-dpdk-parser-state-err.p4-error
@@ -1,0 +1,61 @@
+mov h.MainControlT_hdr_3_tcp; h.tcp;
+<DpdkJmpGreaterStatement>(527393)<DpdkJmpGreaterStatement>(527394)jmp LABEL_TRUE_2
+LABEL_TRUE_2
+mov m.MainControlT_tmp_1; 1
+jmp LABEL_END_2
+LABEL_FALSE_2
+mov m.MainControlT_tmp_1; 0
+LABEL_END_2
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_1;
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov m.learnArg; 0
+learn next_hop
+<DpdkJmpGreaterStatement>(527470)<DpdkJmpGreaterStatement>(527471)jmp LABEL_TRUE_3
+LABEL_TRUE_3
+mov m.MainControlT_tmp_2; 1
+jmp LABEL_END_3
+LABEL_FALSE_3
+mov m.MainControlT_tmp_2; 0
+LABEL_END_3
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_2;
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov h.ipv4.srcAddr; t.newAddr;
+mov m.MainControlT_tmp; 0
+mov m.MainControlT_tmp_0; 1234
+learn next_hop
+<DpdkRxStatement>(527672)MAINPARSERIMPL_START
+extract h.tcp;
+<DpdkJmpGreaterStatement>(527183)<DpdkJmpGreaterStatement>(527184)jmp LABEL_TRUE
+LABEL_TRUE
+mov m.MainParserT_parser_tmp; 1
+jmp LABEL_END
+LABEL_FALSE
+mov m.MainParserT_parser_tmp; 0
+LABEL_END
+mov m.local_metadata_rng_result1; m.MainParserT_parser_tmp;
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+<DpdkJmpGreaterStatement>(527312)<DpdkJmpGreaterStatement>(527313)jmp LABEL_TRUE_0
+LABEL_TRUE_0
+mov m.MainControlT_tmp_3; 1
+jmp LABEL_END_0
+LABEL_FALSE_0
+mov m.MainControlT_tmp_3; 0
+LABEL_END_0
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_3;
+jmpnv LABEL_FALSE_1 h.ipv4;
+LABEL_TRUE_1
+apply ipv4_da
+apply ipv4_da2
+jmp LABEL_END_1
+LABEL_FALSE_1
+LABEL_END_1
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(527678)
