+++ testdata/p4_16_samples_outputs/pna-dpdk-table-key-use-annon.p4-error
@@ -1,0 +1,27 @@
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov m.learnArg; 0
+learn next_hop
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov h.ipv4.srcAddr; t.newAddr;
+mov m.MainControlT_tmp; 0
+mov m.MainControlT_tmp_0; 1234
+learn next_hop
+<DpdkRxStatement>(434711)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+jmpnv LABEL_FALSE h.ipv4;
+LABEL_TRUE
+apply ipv4_da
+apply ipv4_da2
+jmp LABEL_END
+LABEL_FALSE
+LABEL_END
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(434717)
