+++ testdata/p4_16_samples_outputs/pna-example-SelectByDirection2.p4-error
@@ -1,0 +1,39 @@
+mov m.local_metadata_meta; t.addr;
+jmpeq LABEL_TRUE_2 m.pna_main_input_metadata_direction; 0
+mov m.MainControlT_addr; h.ipv4.dstAddr;
+jmp LABEL_END_2
+LABEL_TRUE_2
+mov m.MainControlT_addr; h.ipv4.srcAddr;
+LABEL_END_2
+mov m.local_metadata_meta; m.MainControlT_addr;
+drop
+<DpdkRxStatement>(384937)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+jmpnv LABEL_FALSE h.ipv4;
+LABEL_TRUE
+jmpeq LABEL_TRUE_0 m.pna_main_input_metadata_direction; 0
+mov m.MainControlT_addr; h.ipv4.dstAddr;
+jmp LABEL_END_0
+LABEL_TRUE_0
+mov m.MainControlT_addr; h.ipv4.srcAddr;
+LABEL_END_0
+mov m.local_metadata_meta; m.MainControlT_addr;
+jmpneq LABEL_FALSE_0 m.local_metadata_meta; h.ipv4.dstAddr;
+LABEL_TRUE_1
+apply ipv4_da_lpm
+jmp LABEL_END_1
+LABEL_FALSE_0
+LABEL_END_1
+jmp LABEL_END
+LABEL_FALSE
+LABEL_END
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(384943)
