+++ testdata/p4_16_samples_outputs/pna-example-SelectByDirection2.p4.spec
@@ -36,7 +36,7 @@ metadata instanceof main_metadata_t
 header ethernet instanceof ethernet_t
 header ipv4 instanceof ipv4_t
 
-regarray direction_port_mask size 0x1 initval 0
+regarray network_port_mask size 0x1 initval 0
 
 action forward args instanceof forward_arg_t {
 	mov m.local_metadata_meta t.addr
@@ -68,7 +68,7 @@ apply {
 	jmp MAINPARSERIMPL_ACCEPT
 	MAINPARSERIMPL_PARSE_IPV4 :	extract h.ipv4
 	MAINPARSERIMPL_ACCEPT :	jmpnv LABEL_END h.ipv4
-	regrd m.reg_read_tmp direction_port_mask 0x0
+	regrd m.reg_read_tmp network_port_mask 0x0
 	mov m.left_shift_tmp 0x1
 	shl m.left_shift_tmp m.pna_main_input_metadata_input_port
 	mov m.pna_main_input_metadata_direction m.reg_read_tmp
