+++ testdata/p4_16_samples_outputs/pna-example-dpdk-varbit.p4-error
@@ -1,0 +1,30 @@
+mov h.ethernet.dstAddr; t.param;
+mov h.ethernet.etherType; t.param;
+<DpdkRxStatement>(434766)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4_base;
+jmpeq MAINPARSERIMPL_ACCEPT h.ipv4_base.version_ihl; 69
+jmp MAINPARSERIMPL_PARSE_IPV4_OPTIONS
+MAINPARSERIMPL_PARSE_IPV4_OPTIONS
+lookahead h.MainParserT_parser_lookahead_tmp;
+mov m.MainParserT_parser_tmp_2; h.MainParserT_parser_lookahead_tmp.f;
+jmpeq MAINPARSERIMPL_PARSE_IPV4_OPTION_TIMESTAMP m.MainParserT_parser_tmp_2; 68
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4_OPTION_TIMESTAMP
+lookahead h.MainParserT_parser_tmp_hdr;
+<DpdkCastStatement>(434471)mov m.MainParserT_parser_tmp_0; m.MainParserT_parser_tmp;
+<DpdkShlStatement>(434473)mov m.MainParserT_parser_tmp_1; m.MainParserT_parser_tmp_0;
+<DpdkAddStatement>(434475)mov m.MainParserT_parser_tmp_1_extract_tmp; m.MainParserT_parser_tmp_1;
+<DpdkShrStatement>(434537)extract h.ipv4_option_timestamp;m.MainParserT_parser_tmp_1_extract_tmp;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+mov m.pna_main_output_metadata_output_port; 0
+apply tbl
+apply tbl2
+emit h.ethernet;
+emit h.ipv4_base;
+<DpdkTxStatement>(434772)
