+++ testdata/p4_16_samples_outputs/pna-example-mirror-packet.p4-error
@@ -1,0 +1,21 @@
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov m.mirrorSlot; 3
+mov m.mirrorSession; 58
+mirror m.mirrorSlot; m.mirrorSession;
+drop
+mov m.mirrorSlot_0; 3
+mov m.mirrorSession_0; 62
+mirror m.mirrorSlot_0; m.mirrorSession_0;
+<DpdkRxStatement>(395031)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+apply flowTable
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(395037)
