+++ testdata/p4_16_samples_outputs/pna-example-mirror-packet.p4.spec
@@ -35,7 +35,7 @@ struct main_metadata_t {
 }
 metadata instanceof main_metadata_t
 
-regarray direction_port_mask size 0x1 initval 0
+regarray network_port_mask size 0x1 initval 0
 
 action NoAction args none {
 	return
