+++ testdata/p4_16_samples_outputs/pna-example-template.p4-error
@@ -1,0 +1,20 @@
+mov m.pna_main_output_metadata_output_port; t.vport;
+drop
+<DpdkRxStatement>(372111)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+jmpnv LABEL_FALSE h.ipv4;
+LABEL_TRUE
+apply ipv4_da_lpm
+jmp LABEL_END
+LABEL_FALSE
+LABEL_END
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(372117)
