+++ testdata/p4_16_samples_outputs/pna-example-tunnel.p4-error
@@ -1,1 +1,26 @@
 [--Wwarn=mismatch] warning: Mismatched header/metadata struct for key elements in table tunnel_decap_ipv4_tunnel_term_table. Copying all match fields to metadata
+mov m.local_metadata__tunnel_id1; t.tunnel_id;
+mov m.local_metadata__outer_ipv4_dst0; t.dst_addr;
+<DpdkRxStatement>(502996)PACKET_PARSER_START
+extract h.outer_ethernet;
+jmpeq PACKET_PARSER_PARSE_IPV4_OTR h.outer_ethernet.ether_type; 2048
+jmp PACKET_PARSER_ACCEPT
+PACKET_PARSER_PARSE_IPV4_OTR
+extract h.outer_ipv4;
+jmp PACKET_PARSER_ACCEPT
+PACKET_PARSER_REJECT
+PACKET_PARSER_ACCEPT
+jmpneq LABEL_FALSE m.pna_main_input_metadata_direction; 0
+LABEL_TRUE
+mov m.main_control_tunnel_decap_ipv4_tunnel_term_table_outer_ipv4_src_addr; h.outer_ipv4.src_addr;
+mov m.main_control_tunnel_decap_ipv4_tunnel_term_table_outer_ipv4_dst_addr; h.outer_ipv4.dst_addr;
+apply tunnel_decap_ipv4_tunnel_term_table
+jmp LABEL_END
+LABEL_FALSE
+apply tunnel_encap_set_tunnel_encap
+LABEL_END
+emit h.outer_ethernet;
+emit h.outer_ipv4;
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(503002)
