+++ testdata/p4_16_samples_outputs/pna-issue3041.p4-error
@@ -1,0 +1,29 @@
+mov h.ethernet.dstAddr; t.param;
+mov h.ethernet.etherType; t.param;
+<DpdkRxStatement>(420382)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4_base;
+mov m.MainParserT_parser_tmp; h.ipv4_base.version_ihl;
+<DpdkShrStatement>(420041)<DpdkCastStatement>(420042)jmpeq MAINPARSERIMPL_ACCEPT m.MainParserT_parser_tmp_2; 5
+jmp MAINPARSERIMPL_PARSE_IPV4_OPTIONS
+MAINPARSERIMPL_PARSE_IPV4_OPTIONS
+lookahead h.option;
+jmpeq MAINPARSERIMPL_PARSE_IPV4_OPTION_TIMESTAMP h.option.type; 68
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4_OPTION_TIMESTAMP
+<DpdkCastStatement>(420087)mov m.MainParserT_parser_tmp_1; m.MainParserT_parser_tmp_0;
+<DpdkShlStatement>(420089)mov m.MainParserT_parser_tmp_3; m.MainParserT_parser_tmp_1;
+<DpdkAddStatement>(420091)mov m.MainParserT_parser_tmp_3_extract_tmp; m.MainParserT_parser_tmp_3;
+<DpdkShrStatement>(420153)extract h.ipv4_option_timestamp;m.MainParserT_parser_tmp_3_extract_tmp;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+mov m.pna_main_output_metadata_output_port; 0
+apply tbl
+apply tbl2
+emit h.ethernet;
+emit h.ipv4_base;
+<DpdkTxStatement>(420388)
