+++ testdata/p4_16_samples_outputs/pna-lookahead-structure-bit-field.p4-error
@@ -1,0 +1,15 @@
+<DpdkRxStatement>(309500)MAINPARSERIMPL_START
+lookahead h.MainParserT_parser_lookahead_tmp;
+mov m.local_metadata_f1; h.MainParserT_parser_lookahead_tmp.f;
+jmpeq MAINPARSERIMPL_PARSE_H1 m.local_metadata_f1; 1
+jmpeq MAINPARSERIMPL_PARSE_H2 m.local_metadata_f1; 2
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_H2
+extract h.h2;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_H1
+extract h.h1;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+<DpdkTxStatement>(309506)
