+++ testdata/p4_16_samples_outputs/pna-lookahead-structure.p4-error
@@ -1,0 +1,19 @@
+<DpdkRxStatement>(333961)MAINPARSERIMPL_START
+lookahead h.MainParserT_parser_lookahead_tmp_0;
+mov m.MainParserT_parser_tmp_3; h.MainParserT_parser_lookahead_tmp_0.f;
+mov m.MainParserT_parser_tmp_0; m.MainParserT_parser_tmp_3;
+<DpdkShrStatement>(333779)<DpdkCastStatement>(333780)jmpeq MAINPARSERIMPL_PARSE_H1 m.MainParserT_parser_tmp_2; 4660
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_H1
+extract h.h1;
+lookahead h.MainParserT_parser_lookahead_tmp;
+mov m.MainParserT_parser_tmp_4; h.MainParserT_parser_lookahead_tmp.f;
+mov m.MainParserT_parser_tmp; m.MainParserT_parser_tmp_4;
+<DpdkShrStatement>(333869)<DpdkCastStatement>(333870)<DpdkCastStatement>(333872)<DpdkCastStatement>(333874)jmpeq MAINPARSERIMPL_PARSE_H2 m.MainParserT_parser_tmp_1; 1
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_H2
+extract h.h2;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+<DpdkTxStatement>(333967)
