+++ testdata/p4_16_samples_outputs/pna-mux-dismantle.p4-error
@@ -1,0 +1,59 @@
+mov h.MainControlT_hdr_3_tcp; h.tcp;
+<DpdkJmpGreaterStatement>(530864)<DpdkJmpGreaterStatement>(530865)jmp LABEL_TRUE_1
+LABEL_TRUE_1
+mov m.MainControlT_tmp_3; 1
+jmp LABEL_END_1
+LABEL_FALSE_1
+mov m.MainControlT_tmp_3; 0
+LABEL_END_1
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_3;
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov m.learnArg; 0
+learn next_hop
+<DpdkJmpGreaterStatement>(530941)<DpdkJmpGreaterStatement>(530942)jmp LABEL_TRUE_2
+LABEL_TRUE_2
+<DpdkJmpGreaterStatement>(530946)<DpdkJmpGreaterStatement>(530947)jmp LABEL_TRUE_3
+LABEL_TRUE_3
+mov m.MainControlT_tmp_2; m.local_metadata_val1;
+jmp LABEL_END_3
+LABEL_FALSE_3
+mov m.MainControlT_tmp_2; m.local_metadata_val2;
+LABEL_END_3
+mov m.MainControlT_tmp_1; m.MainControlT_tmp_2;
+jmp LABEL_END_2
+LABEL_FALSE_2
+mov m.MainControlT_tmp_1; m.local_metadata_val1;
+LABEL_END_2
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_1;
+mov m.pna_main_output_metadata_output_port; t.vport;
+mov h.ipv4.srcAddr; t.newAddr;
+mov m.MainControlT_tmp; 0
+mov m.MainControlT_tmp_0; 1234
+learn next_hop
+<DpdkRxStatement>(531153)MAINPARSERIMPL_START
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+<DpdkJmpGreaterStatement>(530783)<DpdkJmpGreaterStatement>(530784)jmp LABEL_TRUE
+LABEL_TRUE
+mov m.MainControlT_tmp_4; 1
+jmp LABEL_END
+LABEL_FALSE
+mov m.MainControlT_tmp_4; 0
+LABEL_END
+mov m.local_metadata_rng_result1; m.MainControlT_tmp_4;
+jmpnv LABEL_FALSE_0 h.ipv4;
+LABEL_TRUE_0
+apply ipv4_da
+apply ipv4_da2
+jmp LABEL_END_0
+LABEL_FALSE_0
+LABEL_END_0
+emit h.ethernet;
+emit h.ipv4;
+<DpdkTxStatement>(531159)
