+++ testdata/p4_16_samples_outputs/pna-mux-dismantle.p4.spec
@@ -69,7 +69,7 @@ header ipv4 instanceof ipv4_t
 header tcp instanceof tcp_t
 header MainControlT_hdr_3_tcp instanceof tcp_t
 
-regarray direction_port_mask size 0x1 initval 0
+regarray network_port_mask size 0x1 initval 0
 
 action do_range_checks_1 args instanceof do_range_checks_1_arg_t {
 	mov h.MainControlT_hdr_3_tcp h.tcp
