+++ testdata/p4_16_samples_outputs/pna-subparser.p4-error
@@ -1,0 +1,14 @@
+<DpdkRxStatement>(319476)MAINPARSERIMPL_START
+invalidate h.ethernet;
+invalidate h.ipv4;
+extract h.ethernet;
+jmpeq MAINPARSERIMPL_COMMONPARSER_PARSE_IPV4 h.ethernet.etherType; 2048
+jmp MAINPARSERIMPL_START_0
+MAINPARSERIMPL_COMMONPARSER_PARSE_IPV4
+extract h.ipv4;
+jmp MAINPARSERIMPL_START_0
+MAINPARSERIMPL_START_0
+jmp MAINPARSERIMPL_ACCEPT
+MAINPARSERIMPL_REJECT
+MAINPARSERIMPL_ACCEPT
+<DpdkTxStatement>(319482)
